Muchas mejoras en el código Añadidas mayoria de encantamientos y accesorios de fargos

+++ Content/Projectiles/HeldItem/BloodRayHeldProj.cs
@@ -41,58 +41,62 @@ namespace RemnantOfTheAncientsMod.Content.Projectiles.HeldItem
         public override void AI()
 		{
             Player player = Main.player[Main.myPlayer];
-            Vector2 SubVelocity = new Vector2(Projectile.ai[0], Projectile.ai[1]);
-
-            if (++speed >= 10)
-			{
-                if (rotation++ >= 360)
-                {
-                    rotation = 0;
-                }
-                speed = 0;
-			}
-            Charge++;
-            //Main.NewText(Charge);
-            if (Charge >= 180 /Projectile.ai[2])
+            if (player.whoAmI == Main.myPlayer)
             {
-                float numberProjectiles = 2;
-                float rotation = MathHelper.ToRadians(5);
+                Vector2 SubVelocity = new Vector2(Projectile.ai[0], Projectile.ai[1]);
 
-          
-                Projectile.position += Vector2.Normalize(Projectile.velocity) * 1f;
-                //if (Main.rand.NextFloat() <= (float)1 / 10)
-                //{
-                for (int i = 0; i < numberProjectiles; i++)
+                if (++speed >= 10)
                 {
-                    Vector2 perturbedSpeed = SubVelocity.RotatedBy(MathHelper.Lerp(-rotation, rotation, i / (numberProjectiles - 1))) * .2f;
-                    int p = Projectile.NewProjectile(Projectile.GetSource_FromAI(), player.position, perturbedSpeed * 10, ModContent.ProjectileType<CrimsomBolt>(), (int)(player.HeldItem.damage * 1.5f), player.HeldItem.knockBack, player.whoAmI);
-                    Main.projectile[p].stepSpeed = 10f;
-                    Main.projectile[p].GetAlpha(Color.Red);
-                   // Main.NewText("Fuego");
+                    if (rotation++ >= 360)
+                    {
+                        rotation = 0;
+                    }
+                    speed = 0;
                 }
-                Projectile.Kill();
+                Charge++;
+                //Main.NewText(Charge);
+                if (Charge >= 180 / Projectile.ai[2])
+                {
+                    float numberProjectiles = 2;
+                    float rotation = MathHelper.ToRadians(5);
+
+
+                    Projectile.position += Vector2.Normalize(Projectile.velocity) * 1f;
+                    //if (Main.rand.NextFloat() <= (float)1 / 10)
+                    //{
+                    for (int i = 0; i < numberProjectiles; i++)
+                    {
+                        Vector2 perturbedSpeed = SubVelocity.RotatedBy(MathHelper.Lerp(-rotation, rotation, i / (numberProjectiles - 1))) * .2f;
+                        int p = Projectile.NewProjectile(Projectile.GetSource_FromAI(), player.position, perturbedSpeed * 10, ModContent.ProjectileType<CrimsomBolt>(), (int)(player.HeldItem.damage * 1.5f), player.HeldItem.knockBack, player.whoAmI);
+                        Main.projectile[p].stepSpeed = 10f;
+                        Main.projectile[p].GetAlpha(Color.Red);
+                        // Main.NewText("Fuego");
+                    }
+                    Projectile.Kill();
+                }
             }
-           
 
                 base.AI();
         }
         public override bool PreDraw(ref Color lightColor)
 		{
-			
-			Player player = Main.player[Main.myPlayer];
-			var texture = ModContent.Request<Texture2D>("RemnantOfTheAncientsMod/Content/Effects/MagicCircle/MagicCircleCenter_1");
-            var texture2 = ModContent.Request<Texture2D>("RemnantOfTheAncientsMod/Content/Effects/MagicCircle/MagicCircleExterior_3");
-            Vector2 origin = new Vector2(texture.Width() * 0.5f, texture.Height() * 0.5f);//0.5
-            Vector2 origin2 = new Vector2(texture2.Width() * 0.5f, texture2.Height() * 0.5f);//0.5
 
-            if (Projectile.ai[0] <= 170)
-			{
-				Color color = new Color(Color.Red.R, Color.Red.G, Color.Red.B, 20) * fade;
-				Main.spriteBatch.Draw((Texture2D)texture, player.Center - Main.screenPosition, null, color, rotation, origin, 1.2f, SpriteEffects.None, 0f);
+            Player player = Main.player[Main.myPlayer];
+            if (player.whoAmI == Main.myPlayer)
+            {
+                var texture = ModContent.Request<Texture2D>("RemnantOfTheAncientsMod/Content/Effects/MagicCircle/MagicCircleCenter_1");
+                var texture2 = ModContent.Request<Texture2D>("RemnantOfTheAncientsMod/Content/Effects/MagicCircle/MagicCircleExterior_3");
+                Vector2 origin = new Vector2(texture.Width() * 0.5f, texture.Height() * 0.5f);//0.5
+                Vector2 origin2 = new Vector2(texture2.Width() * 0.5f, texture2.Height() * 0.5f);//0.5
 
-                Main.spriteBatch.Draw((Texture2D)texture2, player.Center - Main.screenPosition, null, color, -rotation, origin2, 1.5f, SpriteEffects.None, 0f);
+                if (Projectile.ai[0] <= 170)
+                {
+                    Color color = new Color(Color.Red.R, Color.Red.G, Color.Red.B, 20) * fade;
+                    Main.spriteBatch.Draw((Texture2D)texture, player.Center - Main.screenPosition, null, color, rotation, origin, 1.2f, SpriteEffects.None, 0f);
+
+                    Main.spriteBatch.Draw((Texture2D)texture2, player.Center - Main.screenPosition, null, color, -rotation, origin2, 1.5f, SpriteEffects.None, 0f);
+                }
             }
-
 			return true;
 		}
 	}
